requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
groq>=0.4.0
gTTS>=2.4.0
pydub>=0.25.1
python-dotenv>=1.0.0
python-multipart>=0.0.6
